Move functions and types to its own package
